+++ src/main/java/domain/game/BlackjackGame.java
@@ -1,6 +1,7 @@
 package domain.game;
 
 import domain.card.Card;
+import domain.state.Hit;
 
 public class BlackjackGame {
     public static Hit start(Card card1, Card card2) {
