+++ src/main/java/domain/state/Blackjack.java
@@ -1,21 +1,21 @@
-package domain.game;
+package domain.state;
 
 import domain.card.Card;
 import domain.user.Hand;
 
 import java.util.List;
 
-public class Bust implements State {
+public class Blackjack implements State {
 
     private final Hand hand;
 
-    public Bust(Hand hand) {
+    public Blackjack(Hand hand) {
         this.hand = hand;
     }
 
     @Override
-    public State draw(final Card card) {
-        throw new IllegalStateException("버스트에서는 카드를 뽑을 수 없는 상태");
+    public State draw(Card card) {
+        throw new IllegalStateException("블랙잭에서는 카드를 뽑을 수 없는 상태");
     }
 
     @Override
